make orientation of 3 point

+++ DAY25/LEC25_03.cpp
@@ -16,7 +16,7 @@ int main() {
     using pdd = pair<double, double>;
     
     auto dst = [&](pdd a, pdd b) {
-      return hypot(a.first-b.first, a.second-b.second);
+      return (double)hypot(a.first-b.first, a.second-b.second);
     };
     
     vector<pdd> p(n);
